+++ frontend/src/lib/header.ts
@@ -1,6 +1,0 @@
-
-export interface Header {
-  Key: string
-  Value: string
-}
-